MVP done, added a separate button for the extra point on touch down so TD only gives 6 points instead of the 7 specified in step 4
