+++ public/index.html
@@ -3,6 +3,12 @@
   <head>
     <meta charset="utf-8" />
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link href="%PUBLIC_URL%//apple-touch-icon.png" rel="apple-touch-icon" sizes="180x180">
+    <link href="%PUBLIC_URL%//favicon-32x32.png"    rel="icon"             sizes="32x32" type="image/png">
+    <link href="%PUBLIC_URL%//favicon-16x16.png"    rel="icon"             sizes="16x16" type="image/png">
+    <link href="%PUBLIC_URL%//site.webmanifest"     rel="manifest">
+    <meta name="msapplication-TileColor" content="#da532c">
+    <meta name="theme-color" content="#ffffff">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <!--
@@ -19,7 +25,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>American Football Scoreboard</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
