+++ src/component/BottomRow.js
@@ -1,5 +1,4 @@
 import React from "react";
-import "./App.css";
 
 const BottomRow = () => {
   return (
