+++ src/scss/index.scss
@@ -1,6 +1,6 @@
 @font-face {
   font-family: "Ozone";
-  src: url("./Ozone.ttf");
+  src: url("./font/Ozone.ttf");
   font-weight: 400;
   font-style: normal;
 }
@@ -56,3 +56,5 @@ button:focus {
 button:active {
   transform: scale(0.99);
 }
+
+@import './App';
